% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_two_vec.R
\name{concat_two_vec}
\alias{concat_two_vec}
\title{concatenate two strings}
\usage{
concat_two_vec(vec1, vec2)
}
\arguments{
\item{vec1}{string}

\item{vec2}{string}
}
\value{
string
}
\description{
Create a new string from two existing strings by concatenating them
}
\examples{
concat_two_vec(c(1,2), c(3,4))
}
